Matteo/unimore_orario_corsi/requirements.txt
flask
flask-restful
google-cloud-firestore
gunicorn
wtforms